+++ examples/jest/__tests__/multipleSpecs.spec.js
@@ -39,7 +39,7 @@ describe("Dog's API", () => {
     })
 
     // add expectations
-    it("returns a sucessful body", done => {
+    it("returns a sucessful body", () => {
       return getMeDogs({
         url,
         port,
@@ -47,7 +47,6 @@ describe("Dog's API", () => {
         expect(response.headers["content-type"]).toEqual("application/json")
         expect(response.data).toEqual(EXPECTED_BODY)
         expect(response.status).toEqual(200)
-        done()
       })
     })
   })
@@ -76,7 +75,7 @@ describe("Dog's API", () => {
     })
 
     // add expectations
-    it("returns a sucessful body", done => {
+    it("returns a sucessful body", () => {
       return getMeDogs({
         url,
         port,
@@ -84,7 +83,6 @@ describe("Dog's API", () => {
         expect(response.headers["content-type"]).toEqual("application/json")
         expect(response.data).toEqual(EXPECTED_BODY)
         expect(response.status).toEqual(200)
-        done()
       })
     })
   })