+++ examples/jest/pactTestWrapper.js
@@ -1,9 +1,5 @@
 jasmine.DEFAULT_TIMEOUT_INTERVAL = 30000
 
-beforeAll(done => {
-  provider.setup().then(() => done())
-})
+beforeAll(() => provider.setup())
 
-afterAll(done => {
-  provider.finalize().then(() => done())
-})
+afterAll(() => provider.finalize())