+++ examples/jest/publish.js
@@ -2,7 +2,6 @@ let publisher = require("@pact-foundation/pact-node")
 let path = require("path")
 
 let opts = {
-  providerBaseUrl: "http://localhost:8080",
   pactFilesOrDirs: [path.resolve(process.cwd(), "pacts")],
   pactBroker: "https://localhost:3000",
   pactBrokerUsername: process.env.PACT_USERNAME,
@@ -10,4 +9,4 @@ let opts = {
   consumerVersion: "2.0.0",
 }
 
-publisher.publishPacts(opts).then(() => done())
+publisher.publishPacts(opts)